DiscordBotFixer/requirements.txt
discord.py==2.6.3
python-dotenv==1.1.1
psycopg2-binary==2.9.10